LayoutTests/fast/dom/shadow/style-of-distributed-node-expected.txt

Tests to ensure that styles of distributed nodes are different if their parent styles are different.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".




PASS getColorProperty("child-a") is "rgb(255, 0, 0)"
PASS getColorProperty("child-b") is "rgb(0, 0, 255)"

PASS successfullyParsed is true

TEST COMPLETE
